Consolidate scale actions.
