+++ src/js/model/rules/scales.js
@@ -6,8 +6,9 @@ var dl = require('datalib'),
     model = require('../'),
     lookup = model.lookup,
     store = require('../../store'),
-    addScale = require('../../actions/addScale'),
-    updateScale = require('../../actions/updateScale').updateScale,
+    actions = require('../../actions/scaleActions'),
+    addScale = actions.addScale,
+    updateScale = actions.updateScale,
     getIn = require('../../util/immutable-utils').getIn,
     addScaleToGroup = require('../../actions/ruleActions').addScaleToGroup;
